wip: rough proof-of-concept for running the socket in a child process
